third_party/blink/web_tests/html/marquee/marquee-without-frame-no-crash-expected.txt

PASS window.open("foo", "bar") is null
PASS successfullyParsed is true

TEST COMPLETE

foo 
